Extract clipboard data selection builder to shared utility
The identical ~40-line clipboard-to-TableSelection mapping was duplicated
in both InlineTableDataGridInstance and TableDataDataGridInstance. Pull it
into buildTableSelectionFromClipboardData() in utils.ts so both classes
delegate to a single implementation.
